+++ contracts/access/OwnableInternal.sol
@@ -1,4 +1,4 @@
-pragma ton-solidity = 0.58.1;
+pragma ton-solidity >= 0.58.1;
 
 /// @title This extension is used to add the owner role to the contract. It is used to manage contracts through internal messages.
 abstract contract OwnableInternal {